// src/infra/authorization-client/authorization-api-client/models/authorization-many-references-body-params.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AuthorizationBodyParams } from './authorization-body-params';

/**
 * 
 * @export
 * @interface AuthorizationManyReferencesBodyParams
 */
export interface AuthorizationManyReferencesBodyParams {
    /**
     * List of references to authorize against.
     * @type {Array<AuthorizationBodyParams>}
     * @memberof AuthorizationManyReferencesBodyParams
     */
    'references': Array<AuthorizationBodyParams>;
}